requirements.txt
langchain
langchain-groq
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pypdf
docx2txt
unstructured
sentence-transformers
python-dotenv
pandas
fastapi
uvicorn
jinja2
python-multipart
streamlit>=1.40.0
altair<5
watchdog
pytest
loguru
